test_requirements.txt
pynt
pynt-of-django
sphinx
sphinx_rtd_theme
django-extensions
django-authtools
django-allauth>=0.23.0
django-bootstrap-form
-e git://github.com/django/django-contrib-comments.git@8091284e028d97612301e536d70b7ffee1bc4745#egg=django_comments
django-debug-toolbar
Markdown